requirements.txt
pydantic>=2.0
python-dotenv
graphviz
langsmith
ollama
langchain-nvidia-ai-endpoints